Bug 1214601 - Fixes to gonk-misc r=dhylands

+++ b2g-info/process.cpp
@@ -19,8 +19,9 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <pwd.h>
-#include <sys/stat.h>
 #include <regex.h>
+#include <stdlib.h>
+#include <sys/stat.h>
 
 using namespace std;
 
@@ -325,7 +326,7 @@ Process::ensure_got_meminfo()
   // Android has this pm_memusage interface to get the data we collect here.
   // But collecting the data from smaps isn't hard, and doing it this way
   // doesn't rely on any external code, which is nice.
-  // 
+  //
   // Also, the vsize value I get out of procrank (which uses pm_memusage) is
   // way lower than what I get out of statm (which matches smaps).  I presume
   // that statm is correct here.
@@ -415,7 +416,7 @@ Process::user()
     m_user = pw->pw_name;
   } else {
     char uid[32];
-    snprintf(uid, sizeof(uid), "%lu", st.st_uid);
+    snprintf(uid, sizeof(uid), "%lu", (unsigned long)st.st_uid);
     m_user = uid;
   }
